+++ common/Utils.lua
@@ -10,12 +10,26 @@ local L = PER.Localization
 -- Current module
 local Utils = PER.Modules.Utils
 
+-----------------------
+--- Local Functions ---
+-----------------------
+
+local function PrintChatMessage(color, prefix, msg)
+	DEFAULT_CHAT_FRAME:AddMessage(color:WrapTextInColorCode(prefix .. ": ") .. tostring(msg))
+end
+
 ------------------------
 --- Module Functions ---
 ------------------------
 
 function Utils:PrintMessage(msg)
-	Addon:PrintMessage(msg)
+	PrintChatMessage(NORMAL_FONT_COLOR, addonName, msg)
+end
+
+function Utils:PrintDebug(msg)
+	if PER.Settings.general["debug-mode"] then
+		PrintChatMessage(ORANGE_FONT_COLOR, addonName .. " (Debug)", msg)
+	end
 end
 
 function Utils:IsAccountProfile()
@@ -28,7 +42,9 @@ function Utils:OpenSettingsOnLoading()
 	local characterRealmKey = AWL.Utils:GetCharacterRealmKey()
 
 	if Percursus_Options_v3.profileKeys[characterRealmKey]["open-settings"] then
-		Addon:OpenCategory()
+		if not Addon:OpenCategory() then
+			self:PrintDebug("In combat. The options menu cannot be opened.")
+		end
 
 		Percursus_Options_v3.profileKeys[characterRealmKey]["open-settings"] = false
 	end