+++ data/NamespaceData.lua
@@ -12,8 +12,4 @@ PER.Modules.Utils = PER.Modules.Utils or {}
 
 local AWL = ArcaneWizardLibrary
 
-AWL:NewAddon(addonName, {
-	debugEnabled = function()
-		return PER.Settings.general and PER.Settings.general["debug-mode"]
-	end
-})
+AWL:NewAddon(addonName)