+++ CHANGELOG.md
@@ -2,6 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Added: TOC version for patch 12.1.0 [retail]
-- Removed: TOC version for patch 12.0.5 [retail]
-- Minor code adjustments
+- Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility